fix(plugins): normalize Windows ESM import paths #62980
