+++ src/plugins/sdk-alias.ts
@@ -26,6 +26,32 @@ export function normalizeJitiAliasTargetPath(targetPath: string): string {
   return process.platform === "win32" ? targetPath.replace(/\\/g, "/") : targetPath;
 }
 
+/**
+ * On Windows, Node's ESM loader requires absolute paths to be expressed as file:// URLs.
+ * Raw drive-letter paths like C:\... can be misread as an unsupported URL scheme ("c:").
+ */
+export function toSafeImportPath(specifier: string): string {
+  if (process.platform !== "win32") {
+    return specifier;
+  }
+  if (specifier.startsWith("file://")) {
+    return specifier;
+  }
+  if (path.win32.isAbsolute(specifier)) {
+    const normalizedSpecifier = specifier.replaceAll("\\", "/");
+    if (normalizedSpecifier.startsWith("//")) {
+      return new URL(`file:${encodeURI(normalizedSpecifier)}`).href;
+    }
+    // When test environments spoof win32, they may still pass POSIX absolute paths
+    // (e.g. /tmp/foo). Preserve pathToFileURL-style encoding for that case.
+    if (normalizedSpecifier.startsWith("/")) {
+      return new URL(`file://${encodeURI(normalizedSpecifier)}`).href;
+    }
+    return new URL(`file:///${encodeURI(normalizedSpecifier)}`).href;
+  }
+  return specifier;
+}
+
 function resolveLoaderModulePath(params: LoaderModuleResolveParams = {}): string {
   return params.modulePath ?? fileURLToPath(params.moduleUrl ?? import.meta.url);
 }
